CM-868: menu changes part 2 - renaming files
